Release 0.1-alpha_20211221
The basic login and logout features are now working but still lacks error handling.  The logout also is a dummy endpoint that does nothing and can't even reply to the frontend because of the Spring JWT filter. Still, progress is not negligible, I say.
Also added two batch files to run the frontend and backend from Windows explorer.