Release 0.1-alpha_20211231
Last commit of the year and quite a massive one at that.
Backend for sonata is now mostly working but only for GET and POST on most of the endpoints. PUT and DELETE have not been coded yet. Also, there is a weird behavior regarding the parsing of invalid dates when creating a new event object and needs to be investigated or reworked.
Still, this is quite a milestone and could be the last commit in a while because I can feel that I will be busy this coming days because of work.
Happy new year and may all our dreams come true. Always remember, the world is waiting for you. Godspeed!