Release 0.1-alpha_20220104
Happy New Year!
First commit of the year. I decided to do some work on this project on the last day of my holidays. There are so many changes I can't quantify but basically, the backend for event creation is finished with all the validations needed to prevent invalid inputs. The many to many relationships relying on the event entity has been changed from a generic set to a linked hashset, allowing the sorting of the dependent entities. Work on the frontend has already started with sonata's calendar already having it's core functionalities. The component for new event is ongoing but it may have to work because I'm expecting a heavy workload on my REAL job this 2022.
Even so, the work on this project will continue and I will see to it that this gets finished with hopefully minimal bugs and glitches.