Release 0.1-alpha_20220116
The report generation for events is finally finished. However, it is tied to an http request instead of running as a background job. This is not ideal especially if there are lots of items to be included on the report. But for now, it will suffice. Next in line would be the frontend. That would be the last requirement before creating an official release version.