Release 0.1-alpha_20220121
This particular constellation - Sonata - will be put on hold until further notice. This constellation is just a remaster of an existing and working app that is in use by the team. As such, efforts to improve the app will be diverted in creating a new app, a personal app, that will become a new constellation on Existence. No worries though because I will finish what I started on Sonata. It's just that a new constellation is better to do during this time.

This commit will not really add that much except for the capability to export all events on a spreadsheet. That is a basic functionality without the capability to even filter records for export. This will be improved upon once I return to work on the constellation. But for now, this is good bye.