Release 0.1-alpha_20220308_aux_01
Another small commit. These commits just keep on coming! This just renamed the class EnumUtilOutputConstant to just EnumUtilOutput because, well, enums are supposed to be constants, right? Putting both enum and constant in a name looks redundant. Because the class is renamed, all other classes that use it are also modified. I think this naming convention will be the standard for enums, moving forward.